Move data and convert to ts
